Added authorization and database seeder

+++ app/Http/Controllers/BlogController.php
@@ -33,15 +33,20 @@ class BlogController extends Controller
         return redirect()->route('posts.index')->with('success', 'Post created!');
     }
 
-    // Show form to edit a post
+   
+ // Show form to edit a post
     public function edit(Post $post)
     {
+        $this->authorize('update', $post);
         return view('posts.edit', compact('post'));
     }
 
-    // Update a post
+  
+       // Update a post
     public function update(Request $request, Post $post)
     {
+        $this->authorize('update', $post);
+
         $request->validate([
             'title' => 'required|max:255',
             'content' => 'required',
@@ -53,8 +58,11 @@ class BlogController extends Controller
     }
 
     // Delete a post
-    public function destroy(Post $post)
+
+       public function destroy(Post $post)
     {
+        $this->authorize('delete', $post);
+
         $post->delete();
         return redirect()->route('posts.index')->with('success', 'Post deleted!');
     }
